requirements.txt
flask>=3.0.0
anthropic>=0.25.0
openai-whisper>=20231117
gTTS>=2.5.1
# ffmpeg must be installed separately (see README)
